requirements.txt
discord.py
xmltodict